update pulumi deploy scripts II

+++ deploy.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-# set -e
+set -e
 
 GREEN=$(tput setaf 2)
 RED=$(tput setaf 1)
@@ -13,16 +13,18 @@ source .env
 # Run new AWS docker container only if its not running
 if ! docker ps -a --format '{{.Names}}' | grep ${CONTAINER} >/dev/null; then
 
-    echo "${GREEN} Creating localstack aws services ${NC}"
-    docker run --rm -d --name ${CONTAINER} -p 4566:4566 -p 4510-4559:4510-4559 localstack/localstack
+    if [[ "$ENV" == "local" ]]; then
+        echo "${GREEN} Creating localstack aws services ${NC}"
+        docker run --rm -d --name ${CONTAINER} -p 4566:4566 -p 4510-4559:4510-4559 localstack/localstack
+    fi
 
     echo "${GREEN} Initializing pulumi stacks ${NC}"
 
     # Init common dev stack
-    pulumi stack init dev --cwd packages/infrastructure >/dev/null 2>&1
+    pulumi stack init ${ENV} --cwd packages/infrastructure >/dev/null 2>&1
 
     # Init monitoring dev stack
-    pulumi stack init dev --cwd packages/monitor >/dev/null 2>&1
+    pulumi stack init ${ENV} --cwd packages/monitor >/dev/null 2>&1
 fi
 
 #####################################
@@ -32,10 +34,10 @@ fi
 if [[ "$RESOURCES" == "all" ]]; then
     echo "${GREEN}Deploying common resources for Youtube Partner Program & Syncing service${NC}"
 
-    pulumi down --skip-preview --stack dev --cwd packages/infrastructure
+    pulumi down --skip-preview --stack ${ENV} --cwd packages/infrastructure
 
     # Deploy the infracture resources
-    pulumi up --skip-preview --stack dev --cwd packages/infrastructure
+    pulumi up --skip-preview --stack ${ENV} --cwd packages/infrastructure
 fi
 
 ######################################################
@@ -46,7 +48,7 @@ echo "${GREEN} Building zip archives for Lambda functions ${NC}"
 yarn nx run monitor:build >/dev/null 2>&1
 
 echo "${GREEN} Deploying monitoring & syncing resources Youtube Syncing service ${NC}"
-pulumi down --skip-preview --stack dev --cwd packages/monitor
+pulumi down --skip-preview --stack ${ENV} --cwd packages/monitor
 
 # Deploy the infracture resources
-pulumi up --skip-preview --stack dev --cwd packages/monitor
+pulumi up --skip-preview --stack ${ENV} --cwd packages/monitor
